move rust cli to a separate module
